@@ -10,7 +10,7 @@ description: flutter_go
 version: 1.0.6
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.1.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -23,7 +23,6 @@ dependencies:
   fluro: ^1.3.4
   image_picker: ^0.5.0
   sqflite: ^1.1.5
-  flutter_markdown: ^0.2.0
   url_launcher: ^5.0.2
   # 本地存储、收藏功能
   shared_preferences: ^0.4.3
@@ -42,6 +41,10 @@ dependencies:
   flutter_bloc: ^0.11.1
   bloc: ^0.12.0
   html: ^0.14.0+2
+  markdown: ^2.0.0
+  meta: ^1.0.5
+  string_scanner: ^1.0.0
+  path: ^1.5.1
   flutter_downloader: ^1.1.7
   path_provider: ^1.1.0
   permission_handler: ^3.0.0
@@ -207,6 +210,8 @@ flutter:
   - lib/widgets/themes/Cupertino/CupertinoTabScaffold/demo.dart
   - lib/widgets/themes/Cupertino/CupertinoTabView/demo.dart
   - lib/widgets/themes/Cupertino/CupertinoTimerPicker/demo.dart
+  - lib/page_demo_package/.demo.json
+  - lib/standard_pages/.pages.json
   - assets/app.db
   - assets/images/
   - assets/fonts/
