@@ -1,5 +1,5 @@
 class Api{
-  // static const String BASE_URL = 'http://flutter-go.alibaba.net/';
+//   static const String BASE_URL = 'http://flutter-go.alibaba.net/';
   static const String BASE_URL = 'https://flutter-go.pub/api/';
 
   static const String DO_LOGIN = BASE_URL+'doLogin';//登陆
@@ -15,7 +15,7 @@ class Api{
 
   static const String FEEDBACK = BASE_URL+'auth/feedback';//建议反馈
 
-  static  const String LOTOUT = BASE_URL+'logout';//退出登陆
+//  static  const String LOTOUT = BASE_URL+'logout';//退出登陆
 
   static const String GET_ALL_COLLECTION = BASE_URL+'auth/getAllUserCollection';//获取全部收藏
 
@@ -28,4 +28,6 @@ class Api{
   static const String SET_THEMECOLOR = BASE_URL+'auth/setThemeColor';//设置主题颜色
 
   static const String GET_THEMECOLOR = BASE_URL +'/getThemeColor';//获取主题颜色
+  static const String GET_WIDGET_TREE = 'http://flutter-go.alibaba.net/' + 'getCateList';//获取widget列表树
 }
+
