@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 class WidgetName2Icon {
   static Map<String,dynamic> icons = {
+    "Developer": Icons.developer_mode,
+    "Standard": Icons.pages ,
     "Element":Icons.explicit,
     "Components":Icons.extension,
-    "Themes":Icons.filter_b_and_w,
+    "Theme":Icons.filter_b_and_w,
     "Form":Icons.table_chart,
     "Frame":Icons.aspect_ratio,
     "Media":Icons.subscriptions,
